Change inline math delimiters
Inline equations are now enclosed by single dollars: $inline-eq$. The
usage of single dollars to enclose inline equation has the advantages:

- Same form as LaTeX, turning easy to copy/paste from LaTeX documents;
- Easier to write.

The disadvantage is that this approach is more prone to false positive,
like when one write a monetary expression, like '$20 dollars'. Still
missing a nice way to escape single dollar in expressions like this.